scripts/requirements.txt
google-cloud-bigqueryn~=3.1.0
